// src/app/model/enterprise-info.ts
export class EnterpriseInfo {
  enterpriseId: number;
  enterprisePhone: string;
  manageName: string;
  fullName: string;
  intro: string;
  enterpriseAddress: string;
  enterpriseQq: string;
  sort: number;
  status: number;
  statusReason: string;
  isInfoComplete: number;
  createTime: Date;
  updateTime: Date;
}